Fix compilation error after cherrypicking

+++ nativelib/src/main/resources/scala-native/platform/posix/unwind.c
@@ -22,8 +22,7 @@ int scalanative_unwind_get_proc_name(void *cursor, char *buffer, size_t length,
                              (unw_word_t *)offset);
 }
 
-int scalanative_unwind_get_reg(void *cursor, int regnum,
-                               unsigned long long *valp) {
+int scalanative_unwind_get_reg(void *cursor, int regnum, uintptr_t *valp) {
     return unw_get_reg((unw_cursor_t *)cursor, regnum, (unw_word_t *)valp);
 }
